+++ demo/env.d.ts
@@ -1,0 +1,21 @@
+/// <reference types="vite/client" />
+
+declare module '@svg-maps/taiwan.main' {
+  const map: Map
+  export default map
+}
+
+declare module '@svg-maps/australia' {
+  const map: Map
+  export default map
+}
+
+declare module '@svg-maps/france.regions' {
+  const map: Map
+  export default map
+}
+
+declare module '@svg-maps/usa' {
+  const map: Map
+  export default map
+}