+++ demo/src/app.vue
@@ -1,28 +1,28 @@
+<script setup lang="ts">
+import MapCheckbox from '@/components/MapCheckbox.vue'
+import MapRadio from './components/MapRadio.vue'
+import MapLink from '@/components/MapLink.vue'
+import MapTooltipHeat from '@/components/MapTooltipHeat.vue'
+
+</script>
+
 <template>
-	<section class="examples">
-		<h1 class="examples__title">
-			Examples of <a href="https://www.npmjs.com/package/vue-svg-map">vue-svg-map</a>
-		</h1>
-		<checkbox-map />
-		<radio-map />
-		<link-map />
-		<tooltip-heat-map />
-	</section>
+  <h1 class="title">
+    Examples of <a href="https://www.npmjs.com/package/vue-svg-map">vue-svg-map</a>
+  </h1>
+
+  <main>
+    <MapCheckbox />
+    <MapRadio />
+    <MapLink />
+    <MapTooltipHeat />
+  </main>
 </template>
 
-<script>
-import CheckboxMap from './components/checkbox-map.vue'
-import RadioMap from './components/radio-map.vue'
-import LinkMap from './components/link-map.vue'
-import TooltipHeatMap from './components/tooltip-heat-map.vue'
-
-export default {
-	name: 'App',
-	components: {
-		CheckboxMap,
-		RadioMap,
-		LinkMap,
-		TooltipHeatMap,
-	},
+<style scoped>
+.title {
+  font-size: 2rem;
+  text-align: center;
+  margin-bottom: 5rem;
 }
-</script>
+</style>