+++ src/index.ts
@@ -1,0 +1,7 @@
+import SvgMap from './components/SvgMap.vue'
+import SvgMapCheckbox from './components/SvgMapCheckbox.vue'
+import SvgMapRadio from './components/SvgMapRadio.vue'
+import './style.css'
+
+export { SvgMap, SvgMapCheckbox, SvgMapRadio }
+export * from './types'