+++ tests/components/mocks/map.ts
@@ -1,0 +1,25 @@
+import type { Map } from '../../../src/types';
+
+const fakeMap: Map = {
+	label: 'label',
+	viewBox: 'viewBox',
+	locations: [
+		{
+			name: 'name0',
+			id: 'id0',
+			path: 'path0',
+		},
+		{
+			name: 'name1',
+			id: 'id1',
+			path: 'path1',
+		},
+		{
+			name: 'name2',
+			id: 'id2',
+			path: 'path2',
+		},
+	],
+}
+
+export default fakeMap;