+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-beta]
+Create Vue 3 version with TypeScript
+
+### Breaking Changes:
+- Rename the components to `SvgMap`, `SvgMapCheckbox` and `SvgMapRadio`
+- Refactor the props
+
 ## [1.2.0]
 ### Added
 - Add slots before and after all the locations